drop DefaultQueryRuntimeTable, wip moving
